[refactor] 기본 과제 구조변경
