+++ src/routes/router.js
@@ -1,46 +1,20 @@
 import render from "../core/render";
-import handleChangeProfile from "../pages/handleChangeProfile";
-import handleLogin from "../pages/handleLogin";
-import handleProfile from "../pages/handleProfile";
+import isLogin from "../utils/isLogin";
 import navigate from "./navigate";
+import handleProfile from "../pages/handleProfile";
 
 const router = () => {
+  const path = window.location.pathname;
+
+  if (path === "/profile") {
+    if (!isLogin()) {
+      navigate("/login");
+      return;
+    }
+  }
+
   render(window.location.pathname);
-  window.addEventListener("popstate", () => {
-    render(window.location.pathname);
-    handleProfile();
-  });
-
-  window.addEventListener("click", (e) => {
-    if (e.target.tagName === "A") {
-      e.preventDefault();
-      if (e.target.id === "logout") {
-        localStorage.removeItem("user");
-        navigate("/login");
-      } else {
-        const path = e.target.getAttribute("href");
-        navigate(path);
-        handleProfile();
-      }
-    }
-  });
-
-  window.addEventListener("submit", (e) => {
-    e.preventDefault();
-    const form = e.target;
-    const formData = new FormData(form);
-    const data = Object.fromEntries(formData);
-
-    if (form.id == "login-form") {
-      handleLogin(data);
-    }
-    if (form.id == "profile-form") {
-      handleChangeProfile(data);
-    }
-    handleProfile();
-  });
-
-  window.addEventListener("DOMContentLoaded", () => handleProfile());
+  handleProfile();
 };
 
 export default router;
